Test of scripting [PASS] now is to implement
here is to create 3 buttons of titretion that has script on click it reduce 1 to 0.7, 0.7 to 0.6, 0.6 to 0.5 for the 1st to last button respectively.

and the conical flask of soln B when filling other CF same way.

and there is change in script jst niw for mesh render 1 for
B nad MS2 for cf